requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
google-cloud-vision==3.4.5
openai==1.3.0
transformers==4.35.2
torch==2.1.0
Pillow==10.1.0
requests==2.31.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
tldextract==5.1.1
pydantic==2.5.0